pyproject.toml
[tool.poetry]
name = "linked-in-scraper"
version = "0.1.0"
description = "Trying to get a job? Search smarter not harder."
authors = ["aaronfeingold <ajfeingold88@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
linked-in-scraper = "linked_in_scraper.main:main"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.8"
pandas = "^2.2.3"
python-jobspy = "^1.1.76"
google-api-python-client = "^2.160.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
tenacity = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
